Enforce 64 bit precision on times
While building with SCREENSAVER_A_MAX_TIME_SEC 3600, I got an error:

 warning: integer overflow in expression of type 'int' results in '-694967296'

This should fix that.

+++ src/defaults.c
@@ -20,8 +20,8 @@ const config_t default_config = {
             .screensaver = {
                 .mode = SCREENSAVER_A_MODE,
                 .only_if_inactive = SCREENSAVER_A_ONLY_IF_INACTIVE,
-                .idle_time_us = SCREENSAVER_A_IDLE_TIME_SEC * 1000000,
-                .max_time_us = SCREENSAVER_A_MAX_TIME_SEC * 1000000,
+                .idle_time_us = (uint64_t)SCREENSAVER_A_IDLE_TIME_SEC * 1000000,
+                .max_time_us = (uint64_t)SCREENSAVER_A_MAX_TIME_SEC * 1000000,
             }
         },
     .output[OUTPUT_B] =
@@ -40,8 +40,8 @@ const config_t default_config = {
             .screensaver = {
                 .mode = SCREENSAVER_B_MODE,
                 .only_if_inactive = SCREENSAVER_B_ONLY_IF_INACTIVE,
-                .idle_time_us = SCREENSAVER_B_IDLE_TIME_SEC * 1000000,
-                .max_time_us = SCREENSAVER_B_MAX_TIME_SEC * 1000000,
+                .idle_time_us = (uint64_t)SCREENSAVER_B_IDLE_TIME_SEC * 1000000,
+                .max_time_us = (uint64_t)SCREENSAVER_B_MAX_TIME_SEC * 1000000,
             }
         },
     .enforce_ports = ENFORCE_PORTS,
